Updates   * Cleaned up the input component
