+++ repos/keg-components/src/components/form/input/index.js
@@ -1,0 +1,1 @@
+export * from './input'
